调整皮肤管理相关 UI
